requirements.txt
Brotli==1.0.9
cffi==1.15.1
click==8.1.3
cssselect2==0.6.0
fonttools==4.37.1
ghp-import==2.1.0
html5lib==1.1
importlib-metadata==4.12.0
Jinja2==3.1.2
Markdown==3.3.7
MarkupSafe==2.1.1
mergedeep==1.3.4
mkdocs==1.3.1
packaging==21.3
Pillow==9.2.0
pycparser==2.21
pydyf==0.2.0
pymdown-extensions==9.5
pyparsing==3.0.9
pyphen==0.13.0
python-dateutil==2.8.2
pytest==7.1.3
PyYAML==6.0
pyyaml-env-tag==0.1
six==1.16.0
tinycss2==1.1.1
watchdog==2.1.9
weasyprint==56.1
webencodings==0.5.1
zipp==3.8.1
zopfli==0.2.1
